Block extraction refactor.
