+++ src/blocks/homepage-articles-list/view.js
@@ -1,0 +1,7 @@
+//alert( "hello!");
+
+/**
+ * Style dependencies
+ */
+
+import '../../shared/homepage-articles/view.scss';
