+++ src/shared/homepage-articles/index.js
@@ -1,0 +1,2 @@
+export { default as attributes } from './attributes';
+export { default as HomepageArticlesComponent } from './homepage-articles-component';
