+++ src/shared/homepage-articles/sass/_variables.scss
