+++ src/shared/homepage-articles/view.scss
